+++ frontend/src/container/contracts/components/Section.tsx
